[project @ 2000-04-29 10:06:31 by doogie] Added a NameToPathHash, that implements the following algorithm, similiar to what has been discussed during several dinstall discussions.
db/12345.status -> db/5/4/3/12345.status
db/12.status -> db/s/2/1/12.status

Still left to do is write a converter program to transform the layout.
I've tested that this works on master.  It has code to work with both the old
and new dir layout, so there should be no problem using this in current
setups.  That is, if DBase.pm was used in any of the perl scripts. :|