[project @ 2003-08-18 01:16:14 by cjwatson] Phase 1 of version tracking: CGI scripts only.
pkgreport.cgi now takes 'version', 'dist', and 'arch' arguments. If
'version' is supplied, then bugs open against that version are displayed.
Otherwise, if 'dist' is supplied, then it's expected to be a distribution,
such as stable, testing, or unstable, and bugs open against the versions of
packages in that distribution are displayed. It defaults to unstable at the
moment. 'arch' is largely for completeness because it was easy, and narrows
this down to the version of a package on that particular architecture. It
defaults to i386.

For now, bugs that apparently don't apply to the given version at all
(because they were found in a later version or in a version on a different
branch) are displayed in a separate group of sections at the bottom, "not
applicable to this version".

This is a bit of a hack at the moment in various ways, mostly because the
version data isn't yet stored in the .status files. The new $gVersion*
configuration variables are subject to completely arbitrary change. Until
all of this settles down, the output should look unchanged unless you
supply a 'version' or 'dist' argument.