[project @ 2003-08-22 01:41:54 by cjwatson] Implement .status format version 2.
The new format is RFC822-style (continuation lines not implemented, but
feel free to do so), and it's stored in .db files alongside .status, to
make migration less of a one-way affair. Field names are more or less those
used internally, with a few slight changes (originator => submitter, msgid
=> message-id, keywords => tags, forwarded => forwarded-to, mergedwith =>
merged-with). A Format-Version: field is included (currently always 2) to
make the guts of the format extensible in the future should anyone wish to
do so. New fields, though, may be added without incrementing the
format-version, which is the main point of this change.

Support for reading the old format (henceforth "version 1") has been
removed, per Adam Heath. A new migration tool, debbugs-upgradestatus, is
provided, and must be run after upgrading to this version and before
turning debbugs back on.

I think I've changed everything that referred to .status files. We'll find
out once bugs.debian.org has been running this for a while ...

There is precious little documentation of any of this yet.