Add IO::InnerFile support to Debbugs::Log and use it
Use the inner_file handle in cgi/bugreport and Debbugs::CGI::Bugreport
to avoid having the entire file in memory all of the time

Eventually we will want to use an index to avoid having to read every
single line of the bug log for every single message. Some day. [If
you're reading this, you could make that day sooner!]