Add sensible-utils as dependency and make package Lintian clean
Lintian was erroring on:

    E: debbugs-local: missing-depends-on-sensible-utils sensible-browser [usr/bin/local-debbugs]

Add this missing dependency to get the number of Lintian errors to zero
so that builds and e.g. Salsa-CI passes. Several Lintian warnings still remain.

Having Salsa-CI in the package will make it less likely that new Lintian
errors are introduced, or at least their are caught immediately.